ADD: support to AD3542r and increased sclk
It is supporting Single SPI (Classic),  Dual SPI, and Quad SPI mode
on the same bitstream.

Changed the regmap, the reg DAC_CUSTOM_CTRL needs to assert the
MULTI_IO_MODE field to choose the SPI mode.

Increased the frequency of the sclk from 16.5Mhz to 66Mhz.

Changed the AD3552r documentation to show the architecture changes.

Signed-off-by: <carlos.souza@analog.com>